Compile contract artifacts in test targets (#1585)
- **Refactor starknet-contract compiler**
- **Emit starknet contracts in test targets**
